105 frontend monitor without group changing group crash api

+++ backend/uasmartsignage/src/main/java/deti/uas/uasmartsignage/Services/MonitorService.java
@@ -43,7 +43,13 @@ public class MonitorService {
         monitorById.setGroup(monitor.getGroup());
         Monitor returnMonitor = monitorRepository.save(monitorById);
         if (group.isMadeForMonitor()) {
-            monitorGroupRepository.deleteById(group.getId());
+            if (group.getId() == monitorById.getGroup().getId()){
+                group.setName(monitor.getName());
+                monitorGroupRepository.save(group);
+            }
+            else{
+                monitorGroupRepository.deleteById(group.getId());
+            }
         }
         return returnMonitor;
     }
